feat(hw/smbios): added linux support
